Change dtype of the written signal to meet 16-bit PCM wav format

+++ utils/audio.py
@@ -8,7 +8,7 @@ import matplotlib.pyplot as plt
 def signal_to_wav(s, fs, filename, path):
     """Write the signal s into path/filename.wav with sampling freq fs"""
 
-    write(os.path.join(path, "{}.wav".format(filename)), fs, s)
+    write(os.path.join(path, "{}.wav".format(filename)), fs, s.astype(np.int16))
 
 def wav_to_signal(filename, path):
     """Read data from path/filename.wav, return the signal in time domain and its sampling freq"""
